pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dicom_report"]

[project]
name = "dicom-report"
version = "0.1.0"
description = "Generate PNG reports from ultrasound images of adnexal masses with deep learning predictions"
readme = "README.md"
authors = [
    { name = "Filip Christiansen", email = "filip.christiansen.2@ki.se" },
]
requires-python = ">=3.12"
dependencies = ["pillow"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ipykernel>=6.28.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

